fix(town): cast size_t to uint32_t in count() method for consistency

+++ source/town.h
@@ -62,7 +62,7 @@ public:
 	void clear();
 
 	uint32_t count() const {
-		return towns.size();
+		return static_cast<uint32_t>(towns.size());
 	}
 
 	bool addTown(Town* town);
